fix: Contact is repaired.

+++ src/main/kotlin/com/jecminek/edventure_backend/domain/contact/ContactService.kt
@@ -1,8 +1,10 @@
 package com.jecminek.edventure_backend.domain.contact
 
 import com.jecminek.edventure_backend.domain.user.UserService
+import com.jecminek.edventure_backend.security.JwtTokenUtil
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.web.server.ResponseStatusException
@@ -16,9 +18,19 @@ class ContactService {
     @Autowired
     lateinit var userService: UserService
 
-    fun create(userId: Long, contactRequest: ContactRequest): ContactResponse {
+    @Autowired
+    lateinit var jwtTokenUtil: JwtTokenUtil
+
+    fun create(userId: Long, contactRequest: ContactRequest, httpHeaders: HttpHeaders): ContactResponse {
+        val loggedUserId = jwtTokenUtil.getUserId(httpHeaders)
+
+        if (userId != loggedUserId)
+            throw ResponseStatusException(HttpStatus.FORBIDDEN, "User CANNOT create contact to other user")
+
+        //Check user existence
         userService.findById(userId)
-        return repository.save(contactRequest.convertToEntity()).convertToResponse()
+
+        return repository.save(contactRequest.convertToEntity(loggedUserId)).convertToResponse()
     }
 
     fun getById(contactId: Long): Contact {
@@ -28,7 +40,18 @@ class ContactService {
         )
     }
 
-    fun update(userId: Long, contactId: Long, contactRequest: ContactRequest): ContactResponse {
+    fun update(
+        userId: Long,
+        contactId: Long,
+        contactRequest: ContactRequest,
+        httpHeaders: HttpHeaders
+    ): ContactResponse {
+
+        val loggedUserId = jwtTokenUtil.getUserId(httpHeaders)
+
+        if (userId != loggedUserId)
+            throw ResponseStatusException(HttpStatus.FORBIDDEN, "User CANNOT update contact to other user")
+
         //CHECK USER EXISTENCE
         userService.findById(userId)
 
@@ -38,8 +61,15 @@ class ContactService {
         return repository.save(contact).convertToResponse()
     }
 
-    fun delete(userId: Long, contactId: Long) {
+    fun delete(userId: Long, contactId: Long, httpHeaders: HttpHeaders) {
+        val loggedUserId = jwtTokenUtil.getUserId(httpHeaders)
+
+        if (userId != loggedUserId)
+            throw ResponseStatusException(HttpStatus.FORBIDDEN, "User CANNOT delete contact to other user")
+
+        //CHECK USER EXISTENCE
         userService.findById(userId)
+
         repository.delete(getById(contactId))
     }
 
@@ -54,10 +84,10 @@ class ContactService {
     }
 
 
-    fun ContactRequest.convertToEntity() = Contact(
+    fun ContactRequest.convertToEntity(userId: Long) = Contact(
         value = value,
         contactType = contactType,
-        owner = userService.findById(ownerId)
+        owner = userService.findById(userId)
     )
 }
